requirements.txt
gspread
oauth2client
pandas
dataframe_image
python-telegram-bot==13.15